backend/requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clip @ git+https://github.com/openai/CLIP.git@ded190a052fdf4585bd685cee5bc96e0310d2c93
ecdsa==0.19.1
fastapi==0.135.1
ffmpeg-python==0.2.0
filelock==3.20.0
fsspec==2025.12.0
ftfy==6.3.1
future==1.0.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pgvector==0.4.2
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.3.0
regex==2026.2.28
rsa==4.9.1
setuptools==70.2.0
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.14.0
torch==2.10.0+cpu
torchvision==0.25.0+cpu
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.41.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
